ADDED: Method to generate total rides total fare and fare

+++ Cab-Invoice-Generator/src/main/java/cabinvoicegenrator/Ride.java
@@ -1,10 +1,10 @@
 package cabinvoicegenrator;
 
 public class Ride {
-	public int distance;
+	public double distance;
     public int time;
 
-    public Ride(int distance, int time) {
+    public Ride(double distance, int time) {
         this.distance = distance;
         this.time = time;
     }
